requirements.txt
access==1.1.9
affine==2.4.0
appnope==0.1.4
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.3
boto3==1.28.43
botocore==1.31.43
bqplot==0.12.40
branca==0.6.0
cachetools==5.3.1
certifi==2023.7.22
cfgv==3.4.0
cftime==1.6.2
charset-normalizer==3.2.0
click==8.1.6
click-plugins==1.1.1
cligj==0.7.2
colour==0.1.5
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
deprecation==2.1.0
distlib==0.3.7
earthengine-api==0.1.360
eerepr==0.0.4
esda==2.5.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
factor_analyzer==0.5.0
filelock==3.13.1
Fiona==1.9.4.post1
folium==0.14.0
fonttools==4.41.1
future==0.18.3
GDAL==3.8.5
geemap==0.24.4
geocoder==1.38.1
geopandas==0.14.4
giddy==2.3.5
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
gurobipy==11.0.0
h11==0.14.0
httplib2==0.22.0
identify==2.5.32
idna==3.4
imbalanced-learn==0.11.0
imblearn==0.0
imdlib==0.1.19
inequality==1.0.1
ipyevents==2.0.1
ipyfilechooser==0.6.0
ipykernel==6.25.0
ipyleaflet==0.17.3
ipython==8.14.0
ipytree==0.2.2
ipywidgets==8.0.7
jedi==0.18.2
jenkspy==0.4.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
libpysal==4.9.2
llvmlite==0.42.0
mapclassify==2.6.1
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mgwr==2.2.1
momepy==0.7.0
mpmath==1.3.0
natsort==8.4.0
nest-asyncio==1.5.6
netCDF4==1.6.4
networkx==3.2.1
nodeenv==1.8.0
numba==0.59.0
numdifftools==0.9.41
numpy==1.25.1
oauth2client==4.1.3
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.9.1
plotly==5.15.0
pointpats==2.4.0
pre-commit==3.5.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
PuLP==2.8.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyDrive==1.3.1
Pygments==2.15.1
pyparsing==3.0.9
pyperclip==1.8.2
pyproj==3.6.0
pyshp==2.3.1
PySocks==1.7.1
pytesseract==0.3.10
python-box==7.0.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
quantecon==0.7.1
rasterio==1.3.8
rasterstats==0.19.0
ratelim==0.1.6
requests==2.31.0
rioxarray==0.14.1
rsa==4.9
Rtree==1.2.0
s3transfer==0.6.2
scikit-learn==1.3.1
scipy==1.11.2
scooby==0.7.2
seaborn==0.13.0
segregation==2.5
selenium==4.10.0
semopy==2.3.10
shapely==2.0.1
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.5
spaghetti==1.7.5.post1
spglm==1.1.0
spint==1.0.7
splot==1.1.5.post1
spopt==0.6.0
spreg==1.4.2
spvcm==0.3.0
stack-data==0.6.2
statsmodels==0.14.0
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.2.0
tobler==0.11.2
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
traittypes==0.2.1
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.24.7
wcwidth==0.2.6
widgetsnbextension==4.0.8
wsproto==1.2.0
xarray==2023.7.0
xlrd==2.0.1
xlwt-future==0.8.0
xyzservices==2023.7.0
